+++ src/orlab/__init__.py
@@ -1,5 +1,5 @@
 from ._enums import FlightDataType, FlightEvent, OrLogLevel
-from .core.helper import Helper
+from .core.helper import EventTimes, Helper
 from .core.jiterator import JIterator
 from .core.openrocket_instance import OpenRocketInstance
 from .core.simulation_listener import AbstractSimulationListener
@@ -9,6 +9,7 @@ from .parallel import SimulationPool
 
 __all__ = [
     "AbstractSimulationListener",
+    "EventTimes",
     "FlightDataType",
     "FlightEvent",
     "FlightSummary",